requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pandas
numpy
python-dotenv
PyJWT